Move index structures into spentindex.h
txdb.h does not include validation.h anymore, so we need these structs somewhere else
